mkinitcpio: Allow additional argument to preset options
Sometimes it might be handy to add arguments to the preset options. For
instance if you are testing a new mkinitcpio feature with a preexisting
preset.

This patch allows us to use any arguments passed after "--" as an
extension to the preset options array.

    mkinitcpio -p linux -- --uefi /efi/EFI/Linux/systemd-linux.efi --microcode /boot/intel-ucode.img

Would use "linux.preset" and append the --uefi and --microcode flags to
the preset option arrays.

Signed-off-by: Morten Linderud <morten@linderud.pw>